fix: update the generated provider file
Cause
The ServiceProviderBuilder did not emit an updated version since the `@GenerateServiceProvider` annotation
doesn't exist in the most files.

Solution
We added a new `generatedProviderFile` option to the preflightBuilder configuration. You need to put the relative path
to the generated provider file (`*.catalyst_builder.g.dart`) in this option.
The PreflightBuilder will automatically delete the file if it exists. This lead to a full regeneration of the service
provider file. 🙌